+++ jsonata.js
@@ -41,7 +41,6 @@ var jsonata = (function() {
         '=': 40,
         '<': 40,
         '>': 40,
-        '`': 80,
         '^': 40,
         '**': 60,
         '..': 20,
@@ -247,10 +246,27 @@ var jsonata = (function() {
                     };
                 }
             }
+            // test for quoted names (backticks)
+            var name;
+            if(currentChar === '`') {
+                // scan for closing quote
+                position++;
+                var end = path.indexOf('`', position);
+                if(end !== -1) {
+                    name = path.substring(position, end);
+                    position = end + 1;
+                    return create('name', name);
+                }
+                position = length;
+                throw {
+                    code: "S0105",
+                    stack: (new Error()).stack,
+                    position: position
+                };
+            }
             // test for names
             var i = position;
             var ch;
-            var name;
             for (;;) {
                 ch = path.charAt(i);
                 if (i === length || ' \t\n\r\v'.indexOf(ch) > -1 || operators.hasOwnProperty(ch)) {
@@ -584,15 +600,44 @@ var jsonata = (function() {
     // and builds on the Javascript framework described by Douglas Crockford at http://javascript.crockford.com/tdop/tdop.html
     // and in 'Beautiful Code', edited by Andy Oram and Greg Wilson, Copyright 2007 O'Reilly Media, Inc. 798-0-596-51004-6
 
-    var parser = function (source) {
+    var parser = function (source, recover) {
         var node;
         var lexer;
 
         var symbol_table = {};
+        var errors = [];
+
+        var remainingTokens = function() {
+            var remaining = [];
+            if(node.id !== '(end)') {
+                remaining.push({type: node.type, value: node.value, position: node.position});
+            }
+            var nxt = lexer();
+            while(nxt !== null) {
+                remaining.push(nxt);
+                nxt = lexer();
+            }
+            return remaining;
+        };
 
         var base_symbol = {
             nud: function () {
-                return this;
+                // error - symbol has been invoked as a unary operator
+                var err = {
+                    code: 'S0211',
+                    token: this.value,
+                    position: this.position
+                };
+
+                if(recover) {
+                    err.remaining = remainingTokens();
+                    err.type = 'error';
+                    errors.push(err);
+                    return err;
+                } else {
+                    err.stack = (new Error()).stack;
+                    throw err;
+                }
             }
         };
 
@@ -612,6 +657,22 @@ var jsonata = (function() {
             return s;
         };
 
+        var handleError = function(err) {
+            if(recover) {
+                // tokenize the rest of the buffer and add it to an error token
+                err.remaining = remainingTokens();
+                errors.push(err);
+                var symbol = symbol_table["(error)"];
+                node = Object.create(symbol);
+                node.error = err;
+                node.type = "(error)";
+                return node;
+            } else {
+                err.stack = (new Error()).stack;
+                throw err;
+            }
+        };
+
         var advance = function (id, infix) {
             if (id && node.id !== id) {
                 var code;
@@ -621,13 +682,13 @@ var jsonata = (function() {
                 } else {
                     code = "S0202";
                 }
-                throw {
+                var err = {
                     code: code,
-                    stack: (new Error()).stack,
                     position: node.position,
                     token: node.id,
                     value: id
                 };
+                return handleError(err);
             }
             var next_token = lexer(infix);
             if (next_token === null) {
@@ -646,12 +707,12 @@ var jsonata = (function() {
                 case 'operator':
                     symbol = symbol_table[value];
                     if (!symbol) {
-                        throw {
+                        return handleError( {
                             code: "S0204",
                             stack: (new Error()).stack,
                             position: next_token.position,
                             token: value
-                        };
+                        });
                     }
                     break;
                 case 'string':
@@ -666,12 +727,12 @@ var jsonata = (function() {
                     break;
                     /* istanbul ignore next */
                 default:
-                    throw {
+                    return handleError( {
                         code: "S0205",
                         stack: (new Error()).stack,
                         position: next_token.position,
                         token: value
-                    };
+                    });
             }
 
             node = Object.create(symbol);
@@ -693,6 +754,13 @@ var jsonata = (function() {
                 left = t.led(left);
             }
             return left;
+        };
+
+        var terminal = function(id) {
+            var s = symbol(id, 0);
+            s.nud = function() {
+                return this;
+            };
         };
 
         // match infix operators
@@ -737,10 +805,10 @@ var jsonata = (function() {
             return s;
         };
 
-        symbol("(end)");
-        symbol("(name)");
-        symbol("(literal)");
-        symbol("(regex)");
+        terminal("(end)");
+        terminal("(name)");
+        terminal("(literal)");
+        terminal("(regex)");
         symbol(":");
         symbol(";");
         symbol(",");
@@ -764,9 +832,21 @@ var jsonata = (function() {
         infix("and"); // Boolean AND
         infix("or"); // Boolean OR
         infix("in"); // is member of array
+        terminal("and"); // the 'keywords' can also be used as terminals (field names)
+        terminal("or"); //
+        terminal("in"); //
         infixr(":="); // bind variable
         prefix("-"); // unary numeric negation
         infix("~>"); // function application
+
+        infixr("(error)", 10, function(left) {
+            this.lhs = left;
+
+            this.error = node.error;
+            this.remaining = remainingTokens();
+            this.type = 'error';
+            return this;
+        });
 
         // field wildcard (single level)
         prefix('*', function () {
@@ -806,13 +886,13 @@ var jsonata = (function() {
                 // all of the args must be VARIABLE tokens
                 this.arguments.forEach(function (arg, index) {
                     if (arg.type !== 'variable') {
-                        throw {
+                        return handleError( {
                             code: "S0208",
                             stack: (new Error()).stack,
                             position: arg.position,
                             token: arg.value,
                             value: index + 1
-                        };
+                        });
                     }
                 });
                 this.type = 'lambda';
@@ -836,7 +916,7 @@ var jsonata = (function() {
                     } catch(err) {
                         // insert the position into this error
                         err.position = sigPos + err.offset;
-                        throw err;
+                        return handleError( err );
                     }
                 }
                 // parse the function body
@@ -884,7 +964,7 @@ var jsonata = (function() {
                 }
             }
             advance("]", true);
-            this.lhs = a;
+            this.expressions = a;
             this.type = "unary";
             return this;
         });
@@ -1130,7 +1210,7 @@ var jsonata = (function() {
                     result = {type: expr.type, value: expr.value, position: expr.position};
                     if (expr.value === '[') {
                         // array constructor - process each item
-                        result.lhs = expr.lhs.map(function (item) {
+                        result.expressions = expr.expressions.map(function (item) {
                             return ast_optimize(item);
                         });
                     } else if (expr.value === '{') {
@@ -1197,7 +1277,7 @@ var jsonata = (function() {
                     if (expr.value === 'and' || expr.value === 'or' || expr.value === 'in') {
                         expr.type = 'name';
                         result = ast_optimize(expr);
-                    } else if (expr.value === '?') {
+                    } else /* istanbul ignore else */ if (expr.value === '?') {
                         // partial application
                         result = expr;
                     } else {
@@ -1209,18 +1289,30 @@ var jsonata = (function() {
                         };
                     }
                     break;
+                case 'error':
+                    result = expr;
+                    if(expr.lhs) {
+                        result = ast_optimize(expr.lhs);
+                    }
+                    break;
                 default:
                     var code = "S0206";
                     /* istanbul ignore else */
                     if (expr.id === '(end)') {
                         code = "S0207";
                     }
-                    throw {
+                    var err = {
                         code: code,
-                        stack: (new Error()).stack,
                         position: expr.position,
                         token: expr.value
                     };
+                    if(recover) {
+                        errors.push(err);
+                        return {type: 'error', error: err};
+                    } else {
+                        err.stack = (new Error()).stack;
+                        throw err;
+                    }
             }
             return result;
         };
@@ -1231,14 +1323,18 @@ var jsonata = (function() {
         // parse the tokens
         var expr = expression(0);
         if (node.id !== '(end)') {
-            throw {
+            var err = {
                 code: "S0201",
-                stack: (new Error()).stack,
                 position: node.position,
                 token: node.value
             };
+            handleError(err);
         }
         expr = ast_optimize(expr);
+
+        if(errors.length > 0) {
+            expr.errors = errors;
+        }
 
         return expr;
     };
@@ -1648,8 +1744,8 @@ var jsonata = (function() {
             case '[':
                 // array constructor - evaluate each item
                 result = [];
-                for(var ii = 0; ii < expr.lhs.length; ii++) {
-                    var item = expr.lhs[ii];
+                for(var ii = 0; ii < expr.expressions.length; ii++) {
+                    var item = expr.expressions[ii];
                     var value = yield * evaluate(item, input, environment);
                     if (typeof value !== 'undefined') {
                         if(item.value === '[') {
@@ -3732,6 +3828,28 @@ var jsonata = (function() {
     }
 
     /**
+     * Merges an array of objects into a single object.  Duplicate properties are
+     * overridden by entries later in the array
+     * @param {*} arg - the objects to merge
+     * @returns {*} - the object
+     */
+    function functionMerge(arg) {
+        // undefined inputs always return undefined
+        if(typeof arg === 'undefined') {
+            return undefined;
+        }
+
+        var result = {};
+
+        arg.forEach(function(obj) {
+            for(var prop in obj) {
+                result[prop] = obj[prop];
+            }
+        });
+        return result;
+    }
+
+    /**
      * Reverses the order of items in an array
      * @param {Array} arr - the array to reverse
      * @returns {Array} - the reversed array
@@ -3984,6 +4102,7 @@ var jsonata = (function() {
     staticFrame.bind('append', defineFunction(functionAppend, '<xx:a>'));
     staticFrame.bind('exists', defineFunction(functionExists, '<x:b>'));
     staticFrame.bind('spread', defineFunction(functionSpread, '<x-:a<o>>'));
+    staticFrame.bind('merge', defineFunction(functionMerge, '<a<o>:o>'));
     staticFrame.bind('reverse', defineFunction(functionReverse, '<a:a>'));
     staticFrame.bind('each', defineFunction(functionEach, '<o-f:a>'));
     staticFrame.bind('sort', defineFunction(functionSort, '<af?:a>'));
@@ -4000,20 +4119,23 @@ var jsonata = (function() {
         "S0102": "Number out of range: {{token}}",
         "S0103": "Unsupported escape sequence: \\{{token}}",
         "S0104": "The escape sequence \\u must be followed by 4 hex digits",
-        "S0203": "Expected {{value}} before end of expression",
+        "S0105": "Quoted property name must be terminated with a backquote (`)",
+        "S0201": "Syntax error: {{token}}",
         "S0202": "Expected {{value}}, got {{token}}",
+        "S0203": "Expected {{value}} before end of expression",
         "S0204": "Unknown operator: {{token}}",
         "S0205": "Unexpected token: {{token}}",
+        "S0206": "Unknown expression type: {{token}}",
+        "S0207": "Unexpected end of expression",
         "S0208": "Parameter {{value}} of function definition must be a variable name (start with $)",
         "S0209": "A predicate cannot follow a grouping expression in a step",
         "S0210": "Each step can only have one grouping expression",
-        "S0201": "Syntax error: {{token}}",
-        "S0206": "Unknown expression type: {{token}}",
-        "S0207": "Unexpected end of expression",
+        "S0211": "The symbol {{token}} cannot be used as a unary operator",
         "S0301": "Empty regular expressions are not allowed",
         "S0302": "No terminating / in regular expression",
         "S0402": "Choice groups containing parameterized types are not supported",
         "S0401": "Type parameters can only be applied to functions and arrays",
+        "S0500": "Attempted to evaluate an expression containing syntax error(s)",
         "T0410": "Argument {{index}} of function {{token}} does not match function signature",
         "T0411": "Context value is not a compatible type with argument {{index}} of function {{token}}",
         "T0412": "Argument {{index}} of function {{token}} must be an array of {{type}}",
@@ -4076,20 +4198,43 @@ var jsonata = (function() {
     /**
      * JSONata
      * @param {Object} expr - JSONata expression
+     * @param {boolean} options - recover: attempt to recover on parse error
      * @returns {{evaluate: evaluate, assign: assign}} Evaluated expression
      */
-    function jsonata(expr) {
+    function jsonata(expr, options) {
         var ast;
+        var errors;
         try {
-            ast = parser(expr);
+            ast = parser(expr, options && options.recover);
+            errors = ast.errors;
+            delete ast.errors;
         } catch(err) {
             // insert error message into structure
             err.message = lookupMessage(err);
             throw err;
         }
         var environment = createFrame(staticFrame);
+
+        var timestamp = new Date(); // will be overridden on each call to evalute()
+        environment.bind('now', defineFunction(function() {
+            return timestamp.toJSON();
+        }, '<:s>'));
+        environment.bind('millis', defineFunction(function() {
+            return timestamp.getTime();
+        }, '<:n>'));
+
         return {
             evaluate: function (input, bindings, callback) {
+                // throw if the expression compiled with syntax errors
+                if(typeof errors !== 'undefined') {
+                    var err = {
+                        code: 'S0500',
+                        position: 0
+                    };
+                    err.message = lookupMessage(err);
+                    throw err;
+                }
+
                 if (typeof bindings !== 'undefined') {
                     var exec_env;
                     // the variable bindings have been passed in - create a frame to hold these
@@ -4104,18 +4249,14 @@ var jsonata = (function() {
                 exec_env.bind('$', input);
 
                 // capture the timestamp and put it in the execution environment
-                // the $now() function will return this value - whenever it is called
-                var timestamp = (new Date()).toJSON();
-                exec_env.bind('now', defineFunction(function() {
-                    return timestamp;
-                }, '<:s>'));
+                // the $now() and $millis() functions will return this value - whenever it is called
+                timestamp = new Date();
 
                 var result, it;
                 // if a callback function is supplied, then drive the generator in a promise chain
                 if(typeof callback === 'function') {
                     exec_env.bind('__jsonata_async', true);
                     var thenHandler = function (response) {
-                        //                    console.log('THEN: ', response);
                         result = it.next(response);
                         if (result.done) {
                             callback(null, result.value);
@@ -4152,11 +4293,17 @@ var jsonata = (function() {
             registerFunction: function(name, implementation, signature) {
                 var func = defineFunction(implementation, signature);
                 environment.bind(name, func);
+            },
+            ast: function() {
+                return ast;
+            },
+            errors: function() {
+                return errors;
             }
         };
     }
 
-    jsonata.parser = parser;
+    jsonata.parser = parser; // TODO remove this in a future release - use ast() instead
 
     return jsonata;